backend/requirements.txt
# Core FastAPI and ASGI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Azure SDK dependencies
azure-identity==1.15.0
azure-cognitiveservices-speech==1.34.0
azure-ai-translator==1.0.0
azure-openai==1.0.0
azure-storage-blob==12.19.0
azure-cosmos==4.4.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.0.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pymongo==4.6.0

# Audio processing dependencies
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1

# HTTP and networking
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
marshmallow==3.20.1

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.8
bcrypt==4.1.2

# Caching and session management
redis==5.0.1
aioredis==2.0.1

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Development and code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
tenacity==8.2.3

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# File handling
pathlib2==2.3.7
watchdog==3.0.0

# Environment-specific dependencies
# Uncomment for production deployment
# gunicorn==21.2.0
# uvicorn[standard]==0.24.0
